Move utils to common folder
