+++ public/index.ts
@@ -21,3 +21,4 @@ export function plugin(initializerContext: PluginInitializerContext) {
   return new SecurityPlugin(initializerContext);
 }
 export { SecurityPluginSetup, SecurityPluginStart } from './types';
+export { tenantColumn } from './apps/configuration/utils/tenant-utils';
